fix(app marketplace): prevent app marketplace image reload (#474)
