+++ src/assets/locales/de/main.json
@@ -300,7 +300,7 @@
           "modularproduction": "Modular Production"
         },
         "noMatch": "Keine Treffer",
-        "for": "für"
+        "for": "zu Ihrem Suchausdruck"
       }
     },
     "semantichub": {
