+++ src/components/pages/AdminBoardDetail/components/BoardPrivacy/index.tsx
@@ -22,7 +22,7 @@ import { useTranslation } from 'react-i18next'
 import { Typography } from '@catena-x/portal-shared-components'
 import { uniqueId } from 'lodash'
 import { Apartment, Person, LocationOn, Web, Info } from '@mui/icons-material'
-import type { AppDetails } from 'features/apps/apiSlice'
+import type { AppDetails } from 'features/apps/types'
 import './BoardPrivacy.scss'
 import { PrivacyPolicyType } from 'features/adminBoard/adminBoardApiSlice'
