+++ src/components/pages/AdminBoardDetail/components/BoardTechnicalUserSetup/index.tsx
@@ -21,7 +21,7 @@
 import { useTranslation } from 'react-i18next'
 import { Typography } from '@catena-x/portal-shared-components'
 import { Grid } from '@mui/material'
-import type { AppDetails } from 'features/apps/apiSlice'
+import type { AppDetails } from 'features/apps/types'
 
 export default function BoardTechnicalUserSetup({
   item,
