+++ src/components/pages/AppDetail/components/AppDetailTechUserSetup/index.tsx
@@ -20,7 +20,7 @@
 
 import { useTranslation } from 'react-i18next'
 import { Typography } from '@catena-x/portal-shared-components'
-import type { AppDetails } from 'features/apps/apiSlice'
+import type { AppDetails } from 'features/apps/types'
 import { Grid } from '@mui/material'
 import '../../AppDetail.scss'
