+++ src/components/pages/AppMarketplace/components/HeaderSection/index.tsx
@@ -1,6 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021, 2023 BMW Group AG
- * Copyright (c) 2021, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021, 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,22 +17,21 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { Chip, Typography } from '@catena-x/portal-shared-components'
-import type { AppDetails } from 'features/apps/types'
-import './AppDetailTags.scss'
-import '../../AppDetail.scss'
+import { Typography } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 
-export default function AppDetailTags({ item }: { item: AppDetails }) {
+export default function HeaderSection() {
   const { t } = useTranslation()
 
-  const tags = item.tags
   return (
-    <div id="tags" className="appdetail-tags">
-      <Typography variant="h3"> {t('content.appdetail.tags')}: </Typography>
-      {tags.map((tag, i) => (
-        <Chip key={i} label={tag} withIcon={false} type="plain" />
-      ))}
-    </div>
+    <section style={{ paddingBottom: 0, margin: '0 auto' }}>
+      <Typography
+        sx={{ fontFamily: 'LibreFranklin-Light', fontSize: '25px' }}
+        variant="h3"
+        className="section-title"
+      >
+        {t('content.appstore.appOverviewSection.title')}
+      </Typography>
+    </section>
   )
 }
