+++ src/components/pages/AppOverviewNew/index.tsx
@@ -21,10 +21,8 @@
 import { useTranslation } from 'react-i18next'
 import { PageBreadcrumb } from 'components/shared/frame/PageBreadcrumb/PageBreadcrumb'
 import { ErrorBar, PageHeader } from '@catena-x/portal-shared-components'
-import {
-  type AppMarketplaceApp,
-  useFetchProvidedAppsQuery,
-} from 'features/apps/apiSlice'
+import { type AppMarketplaceApp } from 'features/apps/types'
+import { useFetchProvidedAppsQuery } from 'features/apps/apiSlice'
 import NoItems from '../NoItems'
 import { AppOverviewList } from '../AppOverview/AppOverviewList'
 import { appToCard } from 'features/apps/mapper'
