+++ src/components/pages/Organization/AppSubscriptions.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { SubscriptionStatus, type ImageType } from 'features/apps/apiSlice'
+import { SubscriptionStatus, type ImageType } from 'features/apps/types'
 import { Button, Image, LogoGrayData } from '@catena-x/portal-shared-components'
 import { fetchImageWithToken } from 'services/ImageService'
 import './Organization.scss'
