+++ src/components/shared/basic/LoadingProgress/index.tsx
@@ -9,7 +9,7 @@
  * https://www.apache.org/licenses/LICENSE-2.0.
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * distributed under the License is distributed on an 'AS IS' BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations
  * under the License.
@@ -17,26 +17,17 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import StageSection from './components/StageSection'
-import AppListSection from './components/AppListSection'
-import SearchSection from './components/SearchSection'
-import HeaderSection from './components/HeaderSection'
-import { Box } from '@mui/material'
-import PageService from 'services/PageService'
-import { useRef } from 'react'
-import './AppMarketplace.scss'
+import { CircleProgress } from '@catena-x/portal-shared-components'
 
-export default function AppMarketplace() {
-  const reference = PageService.registerReference('AppList', useRef(null))
-
+export default function LoadingProgress() {
   return (
-    <main className="app-store">
-      <StageSection />
-      <Box ref={reference} className="overview-section">
-        <HeaderSection />
-        <SearchSection />
-        <AppListSection />
-      </Box>
-    </main>
+    <CircleProgress
+      size={40}
+      step={1}
+      interval={0.1}
+      colorVariant={'primary'}
+      variant={'indeterminate'}
+      thickness={8}
+    />
   )
 }
