+++ src/components/shared/templates/Subscription/SubscriptionElements.tsx
@@ -39,7 +39,7 @@ import './Subscription.scss'
 import AppSubscriptionDetailOverlay from 'components/pages/AppSubscription/AppSubscriptionDetailOverlay'
 import ActivateSubscriptionOverlay from 'components/pages/AppSubscription/ActivateSubscriptionOverlay'
 import { useState, useReducer } from 'react'
-import { SubscriptionStatus } from 'features/apps/apiSlice'
+import { SubscriptionStatus } from 'features/apps/types'
 import ActivateServiceSubscription from 'components/overlays/ActivateServiceSubscription'
 import { SubscriptionTypes } from '.'
 import { getAssetBase } from 'services/EnvironmentService'
