+++ src/features/apps/index.ts
@@ -22,8 +22,10 @@ import { combineReducers } from 'redux'
 import { slice as details } from './details/slice'
 import { slice as favorites } from './favorites/slice'
 import { slice as marketplace } from './marketplaceDeprecated/slice'
+import { slice as control } from './control'
 
 export default combineReducers({
+  control: control.reducer,
   details: details.reducer,
   favorites: favorites.reducer,
   marketplace: marketplace.reducer,
