+++ src/features/info/search/mapper.ts
@@ -20,7 +20,7 @@
 
 import type { CardItems } from '@catena-x/portal-shared-components'
 import type { TenantUser } from 'features/admin/userApiSlice'
-import type { AppMarketplaceApp } from 'features/apps/apiSlice'
+import type { AppMarketplaceApp } from 'features/apps/types'
 import type { BusinessPartner } from 'features/partnerNetwork/types'
 import { SearchCategory, type SearchItem } from './types'
